% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fiji-imagej_bunwarpj_reg.R
\name{aspect.calc}
\alias{aspect.calc}
\title{Calculate the direction component of the vector field.}
\usage{
aspect.calc(x, y)
}
\arguments{
\item{x}{x displacement magnitude}

\item{y}{y displacement magnitude}
}
\value{
The aspect of displacement vectors \code{x} and \code{y}.
}
\description{
Calculate the direction component of the vector field
    with a geographic reference system - 0 degrees as due
    north and 90 degrees as due east (i.e. aspect)
}
\examples{
aspect.calc(3,4)
}
