% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fiji-imagej_bunwarpj_reg.R
\name{dem.displacement.mapping}
\alias{dem.displacement.mapping}
\title{3D DEM displacement mapping from BUnwarpJ}
\usage{
dem.displacement.mapping(tx_file, r_source, r_target, is_inverse = FALSE)
}
\arguments{
\item{tx_file}{The raw transformation file (.txt) from BUnwarpJ.}

\item{r_source}{The DEM raster filename used as source for image registration.}

\item{r_target}{The DEM raster filename used as target for image registration.}

\item{is_inverse}{(\code{logical}) \code{TRUE} if image registraiton performed
in the inverse direction, otherwise \code{FALSE}.}
}
\value{
Returns a data frame containing the following columns:
\item{x_source & y_source}{x & y coordinates from a CRS for the registration}
\item{z_source}{the elevation value (intensity) corresponding the source
locations}
\item{x_target & y_target}{x & y coordinates from a CRS representing the x &
y transformation for every location in the source image}
\item{z_target}{the elevation value (intensity) corresponding the target
locations (from the target image /elevation model)}
\item{x_disp}{x displacement of registered image (from source image)}
\item{y_disp}{y displacement of registered image (from source image)}
\item{z_disp}{z displacement of registered image (from source image)
(z_target - z_source)}
\item{xy_disp}{magnitude of the displacement in 2D (xy)}
\item{xyz_disp}{magnitude of the displacement in 3D (xyz)}
\item{aspect}{geographic orientation of vector direction with 0 degrees
being due north and 90 degrees being due east}
\item{slope}{angle in degrees of displacement in z direction}
}
\description{
Perform 3D displacement mapping from DEMs using raw BUnwarpJ transformation file
}
