% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fiji-imagej_bunwarpj_reg.R
\name{displacements.scale}
\alias{displacements.scale}
\title{Scale displacements}
\usage{
displacements.scale(d_tx, scale_factor, xyz_disp = TRUE,
  return_sp = FALSE, r_ref = NA)
}
\arguments{
\item{d_tx}{a data frame containing the transformation vectors imported from bUnwarpJ}

\item{scale_factor}{the scale factor (\code{numeric}) to adjust the displacement magnitudes}

\item{xyz_disp}{(\code{logical}) \code{TRUE} if x,y and z displacements
are to be scaled, and \code{FALSE} if only x,y displacements are to be scaled.}

\item{return_sp}{(\code{logical}) \code{TRUE} to return a
SpatialPointsDataFrame based on the CRS projection of
a reference \code{\link[raster]{raster}} object}

\item{r_ref}{a \code{\link[raster]{raster}} that has the CRS projection
of the displacements}
}
\value{
A data frame of the scaled displacements
}
\description{
Applies a scale factor to adjust the magnitudes of displacement
    vectors
}
