% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fiji-imagej_bunwarpj_reg.R
\name{euclidean.distance}
\alias{euclidean.distance}
\title{Calculate the Euclidean distance between two points.}
\usage{
euclidean.distance(x1, x2, y1, y2)
}
\arguments{
\item{x1}{x coordinate of point 1.}

\item{x2}{x coordinate of point 2.}

\item{y1}{y coordinate of point 1.}

\item{y2}{y coordinate of point 2.}
}
\value{
The Euclidean distance \code{x} and \code{y}.
}
\description{
Calculate the Euclidean distance between two points.
}
\examples{
euclidean.distance(2,3,4,5)
}
