% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fiji-imagej_bunwarpj_reg.R
\name{geotiff.for.imagej}
\alias{geotiff.for.imagej}
\title{GeoTIFF to TIFF format for ImageJ/Fiji}
\usage{
geotiff.for.imagej(file_name, write_name)
}
\arguments{
\item{file_name}{File name of GeoTIFF raster to be used in ImageJ/Fiji.}

\item{write_name}{File name of new GeoTIFF raster to be used in ImageJ/Fiji.}
}
\value{
A georeferenced image file (GeoTIFF) that can be read by ImageJ/Fiji.
}
\description{
ImageJ/Fiji does not always open GeoTIFF files. For example, written from
   ESRI products or the \code{\link[raster]{writeRaster}} function in
   \code{raster()}. This function gets around this issue by
   re-writing a problematic GeoTIFF using \code{rgdal()}.
}
