% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fiji-imagej_bunwarpj_reg.R
\name{landmarks.to.crs}
\alias{landmarks.to.crs}
\title{Define CRS for points from ImageJ/Fiji}
\usage{
landmarks.to.crs(file_name, r)
}
\arguments{
\item{file_name}{File name of text file with point coordinates that was
saved from ImageJ}

\item{r}{A Raster* object}
}
\description{
This function can transform the points saved from ImageJ/Fiji
    in image coordinates (col, row) with origin (1,1)
    to the  coordinate reference system (CRS) cooresponding
    to the image used for marking.
}
