% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fiji-imagej_bunwarpj_reg.R
\name{matrix.to.raster}
\alias{matrix.to.raster}
\title{Matrix to raster}
\usage{
matrix.to.raster(m, r_ref)
}
\arguments{
\item{m}{Matrix.}

\item{r_ref}{A raster object from {raster} used as a reference
for CRS and extent}
}
\value{
a raster object projected in reference CRS and extent
}
\description{
Converts a matrix to a raster object using the CRS and extent of a reference raster
}
