% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fiji-imagej_bunwarpj_reg.R
\name{reverse.orientation}
\alias{reverse.orientation}
\title{Get reverse orientation}
\usage{
reverse.orientation(angle)
}
\arguments{
\item{angle}{an angle value in degrees}
}
\value{
An angle in degrees of the opposite direction.
}
\description{
Calculates direction in the reverse orientation
}
\examples{
reverse.orientation(90)
}
